refactor: align hashtag UI with design tokens and package layering
