+++ mobile/packages/models/lib/models.dart
@@ -17,6 +17,7 @@ export 'src/divine_filter.dart';
 export 'src/dm_conversation.dart';
 export 'src/dm_message.dart';
 export 'src/feed_type.dart';
+export 'src/hashtag_label_normalize.dart';
 export 'src/hashtag_search_result.dart';
 export 'src/home_feed_response.dart';
 export 'src/log_entry.dart';
